rename: Move domain files about todo
