+++ src/main/java/com/ssafy/todolist/controller/TodolistController.java
@@ -1,8 +1,8 @@
 package com.ssafy.todolist.controller;
 
-import com.ssafy.todolist.domain.FindTodosResponse;
-import com.ssafy.todolist.domain.Todo;
-import com.ssafy.todolist.domain.TodoDTO;
+import com.ssafy.todolist.domain.todo.FindTodosResponse;
+import com.ssafy.todolist.domain.todo.Todo;
+import com.ssafy.todolist.domain.todo.TodoDTO;
 import com.ssafy.todolist.service.TodoService;
 import com.ssafy.todolist.util.ResponseMessage;
 import org.springframework.http.HttpStatus;
