+++ src/main/java/com/ssafy/todolist/domain/paging/CursorResponse.java
@@ -1,6 +1,6 @@
 package com.ssafy.todolist.domain.paging;
 
-import com.ssafy.todolist.domain.TodoDTO;
+import com.ssafy.todolist.domain.todo.TodoDTO;
 
 import java.util.List;
