+++ src/main/java/com/ssafy/todolist/domain/todo/TodoDTO.java
@@ -1,4 +1,4 @@
-package com.ssafy.todolist.domain;
+package com.ssafy.todolist.domain.todo;
 
 public record TodoDTO(int id, String content, boolean completed) {
