+++ src/main/java/com/ssafy/todolist/repository/TodoRepository.java
@@ -1,6 +1,6 @@
 package com.ssafy.todolist.repository;
 
-import com.ssafy.todolist.domain.Todo;
+import com.ssafy.todolist.domain.todo.Todo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
